feat: ErrorCode 추가 및 CustomException 적용
