+++ com.flash.base/src/main/java/com/flash/base/exception/ErrorCode.java
@@ -2,34 +2,10 @@ package com.flash.base.exception;
 
 import org.springframework.http.HttpStatus;
 
-public enum ErrorCode {
+public interface ErrorCode {
 
-    // ------ 4xx ------
-    NOT_FOUND(HttpStatus.BAD_REQUEST, "요청사항을 찾지 못했습니다."),
-    INVALID_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다."),
-    FORBIDDEN(HttpStatus.FORBIDDEN, "접근이 거부되었습니다."),
+    HttpStatus getStatus();
 
-    // ------ 5xx ------
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버에 문제가 발생했습니다."),
-
-    //인증, 권한
-    USER_MISMATCH(HttpStatus.FORBIDDEN, "사용자 정보가 일치하지 않습니다."),
-    USER_ROLE_MISMATCH(HttpStatus.FORBIDDEN, "사용자 권한이 일치하지 않습니다."),;
-
-    private final HttpStatus status;
-    private final String message;
-
-    ErrorCode(final HttpStatus status, String message) {
-        this.status = status;
-        this.message = message;
-    }
-
-    public HttpStatus getStatus() {
-        return status;
-    }
-
-    public String getMessage() {
-        return message;
-    }
+    String getMessage();
 
 }
